Backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
daphne==4.2.0
deep-translator==1.11.4
Django==5.1.3
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
fonttools==4.56.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h5py==3.12.1
httplib2==0.22.0
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
incremental==24.7.2
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.0
msgpack==1.1.0
namex==0.0.8
numpy==2.1.3
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pygraphviz==1.14
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
six==1.16.0
sklearn-compat==0.1.3
soupsieve==2.6
sqlparse==0.5.2
threadpoolctl==3.5.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.2
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.2
